RTC: Add SSE4_1 optimization for aom_vector_var
This commit also fixes a potential integer overflow issue in the C code
when the function is applied to BLOCK_128X128.

Performance:
Functional Level:
| Width | C Code | SSE4_1 | Gain |
|-------|--------|--------|------|
|   16  |  42430 |  25017 | 1.70 |
|   32  |  69805 |  31255 | 2.28 |
|   64  | 126035 |  43659 | 2.80 |
|  128  | 238695 |  72269 | 3.30 |

Encoder Level:
| SPD_SET | TESTSET  | AVG_PSNR | OVR_PSNR |  SSIM   | ENC_T |
|---------|----------|----------|----------|---------|-------|
|    7    |   rtc    | +0.000%  | +0.000%  | +0.000% | -1.2% |
|    7    | rtc_derf | +0.000%  | +0.000%  | +0.000% | -1.2% |
|---------|----------|----------|----------|---------|-------|
|    8    |   rtc    | +0.000%  | +0.000%  | +0.000% | -1.2% |
|    8    | rtc_derf | +0.000%  | +0.000%  | +0.000% | -0.8% |
|---------|----------|----------|----------|---------|-------|
|    9    |   rtc    | +0.000%  | +0.000%  | +0.000% | -0.1% |
|    9    | rtc_derf | +0.000%  | +0.000%  | +0.000% | -0.1% |
|---------|----------|----------|----------|---------|-------|
|   10    |   rtc    | +0.000%  | +0.000%  | +0.000% | -0.1% |
|   10    | rtc_derf | +0.000%  | +0.000%  | +0.000% | -0.1% |

+++ aom_dsp/avg.c
@@ -547,6 +547,6 @@ int aom_vector_var_c(const int16_t *ref, const int16_t *src, const int bwl) {
   }
 
   // (mean * mean): dynamic range 31 bits.
-  var = sse - ((mean * mean) >> (bwl + 2));
+  var = sse - (((unsigned int)mean * mean) >> (bwl + 2));
   return var;
 }
